VVViewer少し直し

+++ GgafDx/include/jp/ggaf/dx/model/Model.h
@@ -234,7 +234,7 @@ public:
      */
     void prepareVtx(void* prm_paVtxBuffer, UINT prm_size_of_vtx_unit,
                     Frm::Model3D* model_pModel3D,
-                    uint16_t* paNumVertices);
+                    uint32_t* paNumVertices);
 
     void transformPosNomalVtx(void* prm_paVtxBuffer, UINT prm_size_of_vtx_unit, int prm_vtx_num);
     void transformPosVtx(void* prm_paVtxBuffer, UINT prm_size_of_vtx_unit, int prm_vtx_num);
